requirements.txt
groq==0.37.1
python-dotenv==1.2.1
sentence-transformers==4.1.0
faiss-cpu==1.32.2
PyMuPDF==1.26.7
tqdm==4.67.1